requirements.txt
transformers>=4.44.0
accelerate>=0.20.0
pillow>=10.0.0
requests>=2.31.0
flask>=3.0.0
safetensors>=0.4.0
torch>=2.0.0
gunicorn>=21.2.0
aiohttp>=3.9.0

